+++ src/Emulator/Cores/Arm64/ARMv8A.cs
@@ -23,7 +23,7 @@ namespace Antmicro.Renode.Peripherals.CPU
 {
     public partial class ARMv8A : TranslationCPU, IARMTwoSecurityStatesCPU, IPeripheralRegister<ARM_GenericTimer, NullRegistrationPoint>, ICPUWithAArch64Support
     {
-        public ARMv8A(IMachine machine, string cpuType, ARM_GenericInterruptController genericInterruptController, uint cpuId = 0, Endianess endianness = Endianess.LittleEndian)
+        public ARMv8A(IMachine machine, string cpuType, ARM_GenericInterruptController genericInterruptController, ulong cpuId = 0, Endianess endianness = Endianess.LittleEndian)
                 : base(cpuId, cpuType, machine, endianness, CpuBitness.Bits64)
         {
             Affinity = new Affinity(cpuId);
